obs-x264: Allow repeat-headers parameter to be set
This parameter is currently hard-coded to false and can not be changed.
For mpegts container and allow better compatibility with broadcast
mpegts decoders, it is better to repeat the sps/pps headers regularly.
This commit enables this parameter to be set either by the user or the
service (ex: udp or srt w/ mpegts container).