obs-ffmpeg: Repeat video headers in IDR and bitstream for jim-nvenc
This commit exposes h264config->repeatSPSPPS & h264config->disableSPSPPS;
the two parameters are set by nvenc to true on default.
We set the second in jim-nvenc to false so that SPS/PPS are written into
bitstream.
The first parameter could be omitted but is exposed since repetition of
video headers might not be required.
Both of these parameters allow better decoding with DVB broadcast decoders.